backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-google-genai==2.0.6
langchain-community==0.3.7
langgraph==0.2.53
faiss-cpu==1.9.0
sentence-transformers==3.3.1
python-dotenv==1.0.1
python-multipart==0.0.12
